Fix Postgres round() type error in company_teams avg_score_100
Postgres round(double precision, integer) has no overload; only round(numeric, integer) exists. The avg_score_100 subquery was calling round() directly on a double-precision avg() result, causing a 500 on /api/companies/{id}/teams. Cast to ::numeric before rounding and back to ::float after, matching the existing pattern used elsewhere in this file (e.g. the dimension-average query near the top). Confirmed the real error via a direct query repro in the Render shell: psycopg.errors.UndefinedFunction: function round(double precision, integer) does not exist.

+++ app/api_server.py
@@ -1460,7 +1460,7 @@ def company_teams(company_id: int, request: Request) -> dict[str, Any]:
                   t.contact_name, t.contact_email, t.contact_mobile,
                   (SELECT count(*) FROM respondents r
                      WHERE r.team_id = t.team_id AND r.role='sales_person') AS n_respondents,
-                  (SELECT round(avg((s.cognitive_empathy + s.eq + s.pressure_composure + s.storytelling) / 4.0) * 100, 1)
+                  (SELECT round((avg((s.cognitive_empathy + s.eq + s.pressure_composure + s.storytelling) / 4.0) * 100)::numeric, 1)::float
                      FROM audit_scores s
                      JOIN audits a USING (audit_id)
                      JOIN respondents r2 ON r2.respondent_id = a.respondent_id
